requirements.txt
# LLM APIs
openai>=1.63.2
anthropic>=0.46.0
google-generativeai>=0.8.4

# Web Tools
playwright>=1.50.0
html5lib>=1.1
duckduckgo_search>=7.4.4

# Utilities
python-dotenv>=1.0.0
tqdm>=4.67.1
requests>=2.32.3
httpx>=0.28.1

# New dependencies
langchain-openai>=0.0.5
streamlit>=1.32.0 